backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
cryptography==41.0.3
email-validator==2.0.0.post2
python-dotenv==1.0.0
bcrypt==4.0.1
fastapi-limiter==0.1.5
redis==4.6.0